JAMES-3435 Factorize consistency choices
As choosing SERIAL or QUORUM upon reads is a
shared concern between mailbox and message
